fix some type errors, remove withRouter
